fixes for 64 bit machines

+++ regDevCopy.c
@@ -106,9 +106,9 @@ void regDevCopy(unsigned int dlen, unsigned int nelem, volatile void* src, volat
 {
     /* check alignment */
     unsigned int alignment = (1<<dlen)-1;
-    if (((unsigned int)src  & alignment) == 0 &&
-        ((unsigned int)dest & alignment) == 0 &&
-        (pmask == 0 || ((unsigned int)pmask & alignment) == 0))
+    if (((unsigned long)src  & alignment) == 0 &&
+        ((unsigned long)dest & alignment) == 0 &&
+        (pmask == 0 || ((unsigned long)pmask & alignment) == 0))
     {
         /* handle standard element sizes: 1, 2 ,4, 8 bytes */
         switch (dlen<<2 | (pmask!=0)<<1 | (swap!=0))
